move under cmd
