+++ cmd/skaffold/app/cmd/testdata/init/.gitignore
