+++ cmd/skaffold/app/cmd/testdata/init/hello/Dockerfile
