+++ cmd/skaffold/app/cmd/testdata/init/hello/k8s-pod.yaml
