+++ cmd/skaffold/app/cmd/testdata/init/hello/skaffold.yaml
