+++ cmd/skaffold/app/cmd/testdata/init/ignore-tags/Dockerfile
