+++ cmd/skaffold/app/cmd/testdata/init/ignore-tags/k8s-pod.yaml
