+++ cmd/skaffold/app/cmd/testdata/init/microservices/leeroy-app/Dockerfile
