+++ cmd/skaffold/app/cmd/testdata/init/microservices/leeroy-web/kubernetes/deployment.yaml
