+++ cmd/skaffold/app/cmd/testdata/init/microservices/skaffold.yaml
